+++ mlir2chainer/mlir2chainer/tests/test_depthwise_convolution_2d.py
@@ -18,9 +18,9 @@ def test_depthwise_convolution_2d():
     out_w = 4
     inputs  = [mlir.Node('v0', 'float', (batch, in_ch, in_h, in_w))]
     outputs = [mlir.Node('v1', 'float', (batch,out_ch,out_h, out_w))]
-    W = np.random.randn(ch_mul, in_ch, kh, kw)
-    b = np.random.randn(out_ch)
-        
+    W = np.random.randn(ch_mul, in_ch, kh, kw).astype(np.float32)
+    b = np.random.randn(out_ch).astype(np.float32)
+
     nodes = inputs + outputs
     input_names = [ x.name for x in inputs ]
     output_names = [ x.name for x in outputs ]
@@ -28,7 +28,7 @@ def test_depthwise_convolution_2d():
     result = mlir.MLIR('DepthwiseConvolution2D', 'mlir2chainer_test', 0.1, input_names, output_names, nodes, [function])
     result.dump('depthwise_convolution_2d.mlir')
 
-    x = np.random.randn(batch, in_ch, in_h, in_w)
+    x = np.random.randn(batch, in_ch, in_h, in_w).astype(np.float32)
     ref = function.run(x)
     m = mlir2chainer.ChainerNN('depthwise_convolution_2d.mlir')
     with chainer.using_config('train', False):
@@ -50,9 +50,9 @@ def test_depthwise_convolution_2d_dilate():
     dy, dx = (2, 3)
     inputs  = [mlir.Node('v0', 'float', (batch, in_ch, in_h, in_w))]
     outputs = [mlir.Node('v1', 'float', (batch,out_ch,out_h, out_w))]
-    W = np.random.randn(ch_mul, in_ch, kh, kw)
-    b = np.random.randn(out_ch)
-        
+    W = np.random.randn(ch_mul, in_ch, kh, kw).astype(np.float32)
+    b = np.random.randn(out_ch).astype(np.float32)
+
     nodes = inputs + outputs
     input_names = [ x.name for x in inputs ]
     output_names = [ x.name for x in outputs ]
@@ -60,7 +60,7 @@ def test_depthwise_convolution_2d_dilate():
     result = mlir.MLIR('DepthwiseConvolution2D', 'mlir2chainer_test', 0.1, input_names, output_names, nodes, [function])
     result.dump('depthwise_convolution_2d.mlir')
 
-    x = np.random.randn(batch, in_ch, in_h, in_w)
+    x = np.random.randn(batch, in_ch, in_h, in_w).astype(np.float32)
     ref = function.run(x)
     m = mlir2chainer.ChainerNN('depthwise_convolution_2d.mlir')
     with chainer.using_config('train', False):
